+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 2023.3.0 - Feature release
 ### Added
 - Open change password page directly
+### Fixed
+- Password edit form shows old data
 
 ## 2023.2.2 - Bugfix release
 ### Fixed